requirements.txt
scikit-learn
opencv-python
numpy
pandas
matplotlib
pyngrok
flask
streamlit
librosa
google-cloud-speech